fix(solana): change solana exports
