+++ solana_v1/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@everstake/wallet-sdk-solana",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Solana - Everstake Wallet SDK",
   "publishConfig": {
     "access": "public"
